Add support for APM Mission quirks
Also merged MissionController and MissionEditorController into single
MissionController implementation
