+++ QGCApplication.pro
@@ -228,7 +228,6 @@ HEADERS += \
     src/Joystick/JoystickManager.h \
     src/LogCompressor.h \
     src/MG.h \
-    src/MissionEditor/MissionEditorController.h \
     src/MissionManager/MissionManager.h \
     src/MissionManager/MissionController.h \
     src/QGC.h \
@@ -347,7 +346,6 @@ SOURCES += \
     src/Joystick/JoystickManager.cc \
     src/LogCompressor.cc \
     src/main.cc \
-    src/MissionEditor/MissionEditorController.cc \
     src/MissionManager/MissionManager.cc \
     src/MissionManager/MissionController.cc \
     src/QGC.cc \
